requirements.txt
asgiref==3.8.1
bson==0.5.8
dataclasses==0.6
Django==4.2.16
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.7.0
gunicorn==23.0.0
mongoengine==0.29.1
packaging==24.1
PyJWT==2.9.0
pymongo==3.12.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2024.2
